Replace Rust proto build crates with build script step
