+++ Makefile
@@ -3,8 +3,8 @@
 VALIDATOR_IMAGE := "ghcr.io/peggyjv/sommelier-sommelier:main"
 ORCHESTRATOR_IMAGE := "ghcr.io/peggyjv/gravity-bridge-orchestrator:latest"
 
-build_protos:
-	./build_protos.sh
+go_protos:
+	./build_go_protos.sh
 
 e2e_build_images: e2e_clean_slate
 	@docker pull $(VALIDATOR_IMAGE)
@@ -19,22 +19,22 @@ e2e_clean_slate:
 
 e2e_cork_test: e2e_aave_v2_stablecoin_test e2e_cellar_v1_test e2e_cellar_v2_test e2e_proposal_test
 
-# Because of the way `make` works, using the e2e_clean_slate target twice as prerequisites
-# for the individual tests doesn't work when `e2e_cork_test` runs the test targets in series,
+# Because of the way `make` works, using the e2e_clean_slate as as a prerequisite for
+# the individual tests doesn't work when `e2e_cork_test` runs the test targets in series,
 # so we explicitly call the cleanup script in each test target.
-e2e_aave_v2_stablecoin_test:
+e2e_aave_v2_stablecoin_test: go_protos
 	@./clean_slate.sh
 	@E2E_SKIP_CLEANUP=true integration_tests/integration_tests.test -test.failfast -test.v -test.run IntegrationTestSuite -testify.m TestAaveV2Stablecoin || make -s fail
 
-e2e_cellar_v1_test:
+e2e_cellar_v1_test: go_protos
 	@./clean_slate.sh
 	@E2E_SKIP_CLEANUP=true integration_tests/integration_tests.test -test.failfast -test.v -test.run IntegrationTestSuite -testify.m TestCellarV1 || make -s fail
 
-e2e_cellar_v2_test:
+e2e_cellar_v2_test: go_protos
 	@./clean_slate.sh
 	@E2E_SKIP_CLEANUP=true integration_tests/integration_tests.test -test.failfast -test.v -test.run IntegrationTestSuite -testify.m TestCellarV2 || make -s fail
 
-e2e_proposal_test:
+e2e_proposal_test: go_protos
 	@./clean_slate.sh
 	@E2E_SKIP_CLEANUP=true integration_tests/integration_tests.test -test.failfast -test.v -test.run IntegrationTestSuite -testify.m TestScheduledCorkProposal || make -s fail
