+++ build_go_protos.sh
@@ -3,12 +3,6 @@
 set -e
 
 REPO_ROOT=$(git rev-parse --show-toplevel)
-
-echo Building Rust bindings. Output will be in ./steward_proto_rust/prost
-cd steward_proto_rust_build
-cargo run --release
-cd -
-
 GO_OUT=$REPO_ROOT/steward_proto_go
 PROTO_PATH=$REPO_ROOT/proto
