+++ steward/src/cellars/cellar_v1.rs
@@ -3,16 +3,16 @@
 //! To learn more see https://github.com/PeggyJV/cellar-contracts/blob/main/src/base/Cellar.sol
 use deep_space::Address as CosmosAddress;
 use ethers::{abi::AbiEncode, contract::EthCall, types::Address as EthereumAddress};
-use steward_proto::steward::{
-    cellar_v1::Function as StrategyFunction,
-    cellar_v1_governance::{trust_position::Position, Function as GovernanceFunction},
-};
 use GovernanceFunction::*;
 use StrategyFunction::*;
 
 use crate::{
     abi::cellar_v1::*,
     error::Error,
+    proto::{
+        cellar_v1::Function as StrategyFunction,
+        cellar_v1_governance::{trust_position::Position, Function as GovernanceFunction},
+    },
     utils::{
         convert_exchange, encode_fees_distributor_address, encode_swap_params,
         governance_call_error, sp_call_error, sp_call_parse_address, string_to_u256,
