+++ steward/src/commands/cork_proposal/aave_v2_stablecoin/trust_position.rs
@@ -1,11 +1,16 @@
-use crate::{application::APP, cellars, commands::cork_proposal::print_proposal, prelude::*};
+use crate::{
+    application::APP,
+    cellars,
+    commands::cork_proposal::print_proposal,
+    prelude::*,
+    proto::{
+        aave_v2_stablecoin_governance::{Function, SetTrust},
+        governance_call::Call,
+        AaveV2StablecoinGovernance, GovernanceCall,
+    },
+};
 use abscissa_core::{clap::Parser, Command, Runnable};
 use ethers::types::Address as EthereumAddress;
-use steward_proto::steward::{
-    aave_v2_stablecoin_governance::{Function, SetTrust},
-    governance_call::Call,
-    AaveV2StablecoinGovernance, GovernanceCall,
-};
 
 /// TrustPosition subcommand
 #[derive(Command, Debug, Parser)]
