+++ steward/src/commands/cork_proposal/cellar_v1/set_fees_distributor.rs
@@ -1,11 +1,16 @@
-use crate::{application::APP, cellars, commands::cork_proposal::print_proposal, prelude::*};
+use crate::{
+    application::APP,
+    cellars,
+    commands::cork_proposal::print_proposal,
+    prelude::*,
+    proto::{
+        cellar_v1_governance::{Function, SetFeesDistributor},
+        governance_call::Call,
+        CellarV1Governance, GovernanceCall,
+    },
+};
 use abscissa_core::{clap::Parser, Command, Runnable};
 use deep_space::Address as CosmosAddress;
-use steward_proto::steward::{
-    cellar_v1_governance::{Function, SetFeesDistributor},
-    governance_call::Call,
-    CellarV1Governance, GovernanceCall,
-};
 
 /// Fees Distributor subcommand
 #[derive(Command, Debug, Parser)]
