+++ steward/src/cork.rs
@@ -6,6 +6,7 @@ use crate::{
         ErrorKind::{self, *},
     },
     prelude::APP,
+    proto::{self, schedule_request::CallData::*, ScheduleRequest, ScheduleResponse},
     somm_send,
 };
 use abscissa_core::{
@@ -19,13 +20,6 @@ use lazy_static::lazy_static;
 use sha3::{Digest, Keccak256};
 use somm_proto::cork::Cork;
 use std::time::Duration;
-use steward_proto::{
-    self,
-    steward::{
-        self, schedule_request::CallData::*, ScheduleRequest, ScheduleResponse, StatusRequest,
-        StatusResponse,
-    },
-};
 use tonic::{self, async_trait, Code, Request, Response, Status};
 
 pub mod cache;
@@ -42,7 +36,7 @@ lazy_static! {
 pub struct CorkHandler;
 
 #[async_trait]
-impl steward::contract_call_server::ContractCall for CorkHandler {
+impl proto::contract_call_server::ContractCall for CorkHandler {
     async fn schedule(
         &self,
         request: Request<ScheduleRequest>,
