+++ steward/src/server.rs
@@ -5,7 +5,7 @@ use tonic::transport::{Certificate, Identity, Server, ServerTlsConfig};
 
 pub const DEFAULT_CLIENT_CA: &[u8] = include_bytes!("../../tls/sevenseas_ca.crt");
 // for gRPC reflection
-pub const DESCRIPTOR: &[u8] = include_bytes!("../../steward_proto_rust/src/prost/descriptor.bin");
+pub const FILE_DESCRIPTOR_SET: &[u8] = include_bytes!("gen/proto/descriptor.bin");
 
 pub struct ServerConfig {
     pub tls_config: Option<ServerTlsConfig>,
