--- steward_proto_rust/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "steward_proto"
-version = "3.0.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-prost = "0.7"
-prost-types = "0.7"
-tonic = "0.4.3"
-serde = "1.0"
