--- steward_proto_rust/src/lib.rs
@@ -1,5 +1,0 @@
-#![allow(clippy::all)]
-
-pub mod steward {
-    include!("prost/steward.v3.rs");
-}
