--- steward_proto_rust_build/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "steward_proto_rust_build"
-version = "1.0.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-prost = "0.7"
-prost-build = "0.7"
-tonic-build = "0.4"
-walkdir = "2.3.2"
